+++ src/utils/render.js
@@ -2,7 +2,6 @@ export const RenderPosition = {
   AFTERBEGIN: `afterbegin`,
   BEFOREEND: `beforeend`
 };
-const activities = [`Check-in`, `Sightseeing`, `Restaurant`];
 
 export const createElement = (template) => {
   const newElement = document.createElement(`div`);
@@ -20,18 +19,4 @@ export const render = (container, element, place) => {
       container.append(element);
       break;
   }
-};
-
-const addNullToTime = (time) => {
-  return time > 9 ? time : `0` + time;
-};
-
-export const getDiff = (time) => {
-  const hours = addNullToTime(new Date(time).getUTCHours());
-  const minutes = addNullToTime(new Date(time).getUTCMinutes());
-  return `${hours}H ${minutes}M`;
-};
-
-export const getPrefix = (type) => {
-  return activities.some((it) => it === type) ? `at` : `to`;
 };
